Send `manualCalories`, with value `0`, to avoid a `500` error in the FitBit server.
See related issue opened on FitBit: https://community.fitbit.com/t5/Web-API-Development/same-activity-upload-for-some-users-response-500-internal-server-error/m-p/5660030

If we don't send `manualCalories`, the server responds with a `500` error.

Note that FitBit doesn't seem to use the zero value we send: we see non-zero calories appear in the FitBit app for the treadmil activities.