Extract an interface TreadmillController to control the treadmill.
Identify methods on the ph4_walkingpad Controller class that we were calling, and put them into a `TreadmillController` in the domain layer.

Add an implementation of this interface, `WalkingpadController`, in the interfaceadapters layer.

This is preparation for usage of an injected controller in other parts of the app (future REST api coming...).