From tests, spy the controller and assert method calls on it.
Why? We should have assertions that check that our rest api actually calls calls the controller as expected, and specifically, that our controller does the appropriate conversions between the weird walkingpad units for speed and normal km/h speed.

How?
Create a `Spy` class which logs all calls to all methods in an object.

In the restapi tests, define expected calls to the controller in the scenarios, and assert that these methods were actually called.